+++ performance-tests/config/activation_config.js
@@ -1,0 +1,4 @@
+export const activationConfig = {
+  access_token_url: 'https://api-mcshared.uat.cstar.pagopa.it/auth/token',
+  activation_base: 'https://api-rtp.uat.cstar.pagopa.it/rtp/activation'
+};
